services/api/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "esn-wifi-api"
version = "0.1.0"
description = "ESN WiFi Billing API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.32.0",
  "pydantic-settings>=2.6.0",
  "sqlalchemy[asyncio]>=2.0.36",
  "asyncpg>=0.30.0",
  "alembic>=1.14.0",
  "psycopg2-binary>=2.9.10",
  "celery[redis]>=5.4.0",
  "redis>=5.2.0",
  "python-jose[cryptography]>=3.3.0",
  "bcrypt>=4.0.1",
  "cryptography>=43.0.0",
  "python-multipart>=0.0.17",
  "email-validator>=2.2.0",
  "httpx>=0.28.0",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.8.0",
  "pytest>=8.3.0",
  "pytest-asyncio>=0.24.0",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
testpaths = ["tests"]
